Added UserGuide PowerPoint Template. Added Collated Files
