+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -85,7 +85,6 @@ public class AddressBookParser {
             return new ExitCommand();
 
         case HelpCommand.COMMAND_WORD:
-            new TextToSpeechParser().parse("Showing help");
             return new HelpCommand();
 
         case UndoCommand.COMMAND_WORD:
