+++ src/main/java/seedu/address/logic/parser/EmailCommandParser.java
@@ -7,6 +7,7 @@ import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.EmailCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+//@@author hanselblack
 /**
  * Parses input arguments and creates a new EmailCommand object
  */
