+++ src/main/java/seedu/address/logic/parser/MusicCommandParser.java
@@ -5,6 +5,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 import seedu.address.logic.commands.MusicCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+//@@author hanselblack
 /**
  * Parses input arguments and creates a new MusicCommand object
  */
