+++ src/test/data/XmlUtilTest/validAddressBook.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<!-- @@hanselblack -->
 <addressbook>
     <persons>
         <name>Hans Muster</name>
