+++ src/test/java/seedu/address/logic/commands/RemarkCommandTest.java
@@ -21,6 +21,7 @@ import seedu.address.model.UserPrefs;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.Remark;
 
+//@@author hanselblack
 public class RemarkCommandTest {
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
