+++ src/test/java/seedu/address/logic/parser/RemarkCommandParserTest.java
@@ -12,6 +12,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.RemarkCommand;
 import seedu.address.model.person.Remark;
 
+//@@author hanselblack
 /**
  * tests different boundary values for RemarkCommandParser.
  */
