+++ src/test/java/seedu/address/model/person/RemarkTest.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
-
+//@@author hanselblack
 public class RemarkTest {
 
     @Test
