+++ unused/cmudict-en-us.dict
@@ -1,0 +1,3 @@
+add AE D
+music M Y UW Z IH K
+list L IH S T
