+++ unused/resources/edu/cmu/sphinx/models/en-us/en-us/feat.params
@@ -1,0 +1,12 @@
+-lowerf 130
+-upperf 6800
+-nfilt 25
+-transform dct
+-lifter 22
+-feat 1s_c_d_dd
+-svspec 0-12/13-25/26-38
+-agc none
+-cmn live
+-varnorm no
+-model ptm
+-cmninit 41.00,-5.29,-0.12,5.09,2.48,-4.07,-1.37,-1.78,-5.08,-2.05,-6.45,-1.42,1.17
