+++ unused/resources/edu/cmu/sphinx/models/en-us/en-us/noisedict
@@ -1,0 +1,5 @@
+<s> SIL
+</s> SIL
+<sil> SIL
+[NOISE] +NSN+
+[SPEECH] +SPN+
